# swagger_config.json
{
  "swagger": "2.0",
  "info": {
    "title": "Person API",
    "description": "A simple API for managing person records.",
    "version": "1.0.0"
  },
  "host": "http://localhost:5000",
  "basePath": "/api",
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "tags": [
    {
      "name": "Person API",
      "description": "Endpoints for managing person records."
    }
  ],
  "paths": {
    "/persons": {
      "post": {
        "tags": ["Person API"],
        "summary": "Create a new person.",
        "operationId": "createPerson",
        "parameters": [
          {
            "in": "body",
            "name": "person",
            "description": "The person object to be created.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "John Doe",
                  "description": "The name of the person."
                },
                "age": {
                  "type": "integer",
                  "format": "int32",
                  "example": 30,
                  "description": "The age of the person."
                },
                "address": {
                  "type": "string",
                  "example": "123 Main St",
                  "description": "The address of the person."
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "New person created successfully."
          },
          "400": {
            "description": "Invalid input data or missing parameters."
          },
          "500": {
            "description": "Something went wrong while creating the person."
          }
        }
      }
    },
    "/persons/{identifier}": {
      "get": {
        "tags": ["Person API"],
        "summary": "Fetch details of a person by ID or name.",
        "operationId": "getPerson",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The ID or name of the person to fetch."
          }
        ],
        "responses": {
          "200": {
            "description": "Person details retrieved successfully.",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer",
                  "format": "int32",
                  "example": 1,
                  "description": "The ID of the person."
                },
                "name": {
                  "type": "string",
                  "example": "John Doe",
                  "description": "The name of the person."
                },
                "age": {
                  "type": "integer",
                  "format": "int32",
                  "example": 30,
                  "description": "The age of the person."
                },
                "address": {
                  "type": "string",
                  "example": "123 Main St",
                  "description": "The address of the person."
                }
              }
            }
          },
          "404": {
            "description": "Person not found."
          }
        }
      },
      "put": {
        "tags": ["Person API"],
        "summary": "Update details of a person by ID or name.",
        "operationId": "updatePerson",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The ID or name of the person to update."
          },
          {
            "in": "body",
            "name": "person",
            "description": "The updated person object.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Updated Name",
                  "description": "The updated name of the person."
                },
                "age": {
                  "type": "integer",
                  "format": "int32",
                  "example": 35,
                  "description": "The updated age of the person."
                },
                "address": {
                  "type": "string",
                  "example": "456 Elm St",
                  "description": "The updated address of the person."
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Person updated successfully."
          },
          "400": {
            "description": "Invalid data format or missing parameters."
          },
          "404": {
            "description": "Person not found."
          },
          "500": {
            "description": "Something went wrong while updating the person."
          }
        }
      },
      "delete": {
        "tags": ["Person API"],
        "summary": "Delete a person by ID or name.",
        "operationId": "deletePerson",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The ID or name of the person to delete."
          }
        ],
        "responses": {
          "200": {
            "description": "Person deleted successfully."
          },
          "404": {
            "description": "Person not found."
          },
          "500": {
            "description": "Something went wrong while deleting the person."
          }
        }
      }
    }
  }
}